backend/requirements.txt
fastapi
uvicorn
redis
pymongo
neo4j
cloudevents
pycryptodome
python-jose[cryptography]
pydantic
python-dotenv
